chore: ignore training/ and consolidate recent source changes
- Add training/ to .gitignore so local experiment artifacts stay out of VCS.
- Model: change stopping_time to double (both pxd and model.h) and free
  global_opt/initial_state in reset() for clean re-solve.
- ml.polynomial: add default_theta() matching solver-default predictions.
- constraint.c: minor cleanup and fix factors realloc type to uint64_t.
- tests/test_constraint.c: rework large-expression pre-allocation tests.
- PRD.md: document C-level variable vectors and matrix-variable ops.

+++ cbqs/Model.pxd
@@ -13,7 +13,7 @@ cdef extern from "src/model.h":
 		size_t M;
 		int break_item;
 		int n;
-		int stopping_time;
+		double stopping_time;
 		int stop_val;
 		int depth_look_ahead;
 		int num_workers;
