+++ cbqs/src/model.h
@@ -16,7 +16,7 @@ typedef struct {
 	size_t M;
     int break_item;
 	int n;
-	int stopping_time;
+	double stopping_time;
 	int stop_val;
 	int depth_look_ahead;
 	int num_workers;
